rematar-backend/requirements.txt
Flask==1.1.2
Flask_RESTful==0.3.8
Flask_Cors==3.0.8
Flask_SocketIO==4.3.0
Flask_Migrate==2.5.3
Flask-Script==2.0.6
flask_marshmallow==0.12.0
mysqlclient==1.4.6
marshmallow_sqlalchemy==0.23.0
PyJWT==1.7.1
